+++ editor/editor-core/src/__tests__/unit/plugins/extension/index.tsx
@@ -36,7 +36,6 @@ import {
 } from '@atlaskit/editor-test-helpers/schema-builder';
 
 import sendKeyToPm from '@atlaskit/editor-test-helpers/send-key-to-pm';
-import sleep from '@atlaskit/editor-test-helpers/sleep';
 import { createFakeExtensionProvider } from '@atlaskit/editor-test-helpers/extensions';
 
 import { editExtension } from '../../../../plugins/extension/actions';
@@ -219,9 +218,13 @@ describe('extension', () => {
         const { editorView } = editor(
           doc(bodiedExtension(bodiedExtensionAttrs)(paragraph('te{<>}xt'))),
         );
-        expect(editExtension(null)(editorView.state, editorView.dispatch)).toBe(
-          false,
-        );
+        expect(
+          editExtension(null)(
+            editorView.state,
+            editorView.dispatch,
+            editorView,
+          ),
+        ).toBe(false);
       });
 
       describe('macroProvider', () => {
@@ -236,7 +239,11 @@ describe('extension', () => {
           );
           const provider = await macroProviderPromise;
           expect(
-            editExtension(provider)(editorView.state, editorView.dispatch),
+            editExtension(provider)(
+              editorView.state,
+              editorView.dispatch,
+              editorView,
+            ),
           ).toBe(true);
         });
 
@@ -244,7 +251,11 @@ describe('extension', () => {
           const { editorView } = editor(doc(paragraph('te{<>}xt')));
           const provider = await macroProviderPromise;
           expect(
-            editExtension(provider)(editorView.state, editorView.dispatch),
+            editExtension(provider)(
+              editorView.state,
+              editorView.dispatch,
+              editorView,
+            ),
           ).toBe(false);
         });
 
@@ -253,8 +264,12 @@ describe('extension', () => {
             doc(bodiedExtension(bodiedExtensionAttrs)(paragraph('{<>}'))),
           );
           const provider = await macroProviderPromise;
-          editExtension(provider)(editorView.state, editorView.dispatch);
-          await sleep(0);
+          editExtension(provider)(
+            editorView.state,
+            editorView.dispatch,
+            editorView,
+          );
+          await flushPromises();
           expect(editorView.state.doc).toEqualDocument(
             doc(
               bodiedExtension(bodiedExtensionData[0].attrs)(
@@ -285,8 +300,12 @@ describe('extension', () => {
             new MockMacroProvider(inlineExtensionData[1]),
           );
           const provider = await macroProviderPromise;
-          editExtension(provider)(editorView.state, editorView.dispatch);
-          await sleep(0);
+          editExtension(provider)(
+            editorView.state,
+            editorView.dispatch,
+            editorView,
+          );
+          await flushPromises();
           expect(editorView.state.doc).toEqualDocument(
             doc(
               paragraph(
@@ -320,8 +339,12 @@ describe('extension', () => {
               new MockMacroProvider(inlineExtensionData[1]),
             );
             const provider = await macroProviderPromise;
-            editExtension(provider)(editorView.state, editorView.dispatch);
-            await sleep(0);
+            editExtension(provider)(
+              editorView.state,
+              editorView.dispatch,
+              editorView,
+            );
+            await flushPromises();
             expect(editorView.state.doc).toEqualDocument(
               doc(
                 bodiedExtension(bodiedExtensionAttrs)(
@@ -349,7 +372,11 @@ describe('extension', () => {
 
           const provider = await macroProviderPromise;
           expect(
-            editExtension(provider)(editorView.state, editorView.dispatch),
+            editExtension(provider)(
+              editorView.state,
+              editorView.dispatch,
+              editorView,
+            ),
           ).toBe(true);
 
           expect(nodeWithPos).toBeDefined();
@@ -408,12 +435,12 @@ describe('extension', () => {
             editExtension(null, updateHandlerPromise)(
               editorView.state,
               editorView.dispatch,
+              editorView,
             ),
           ).toBe(true);
           await flushPromises();
 
           expect(updateFn).toBeCalledWith(initialValue);
-          await flushPromises();
 
           expect(
             editorView.state.doc.firstChild!.attrs.parameters.content,
@@ -422,23 +449,24 @@ describe('extension', () => {
 
         it('should scroll into view', async () => {
           // using a mock to be able to capture the passed TR at the end
-          const dispatchMock = jest.fn(editorView.dispatch);
+          const dispatchSpy = jest.spyOn(editorView, 'dispatch');
 
           expect(
             editExtension(null, updateHandlerPromise)(
               editorView.state,
-              dispatchMock,
+              editorView.dispatch,
+              editorView,
             ),
           ).toBe(true);
+          await flushPromises();
 
           expect(updateFn).toBeCalledWith(initialValue);
-          await flushPromises();
 
           expect(
             editorView.state.doc.firstChild!.attrs.parameters.content,
           ).toBe(newContent);
 
-          const dispatchedTR = dispatchMock.mock.calls[0][0];
+          const dispatchedTR = dispatchSpy.mock.calls[0][0];
           expect((dispatchedTR as any).scrolledIntoView).toBeTruthy();
         });
       });
@@ -479,6 +507,7 @@ describe('extension', () => {
           editExtension(provider, updateMethodResolvingMacroParams)(
             editorView.state,
             editorView.dispatch,
+            editorView,
           );
 
           await flushPromises();
@@ -501,6 +530,7 @@ describe('extension', () => {
           editExtension(provider, updateMethodResolvingUndefined)(
             editorView.state,
             editorView.dispatch,
+            editorView,
           );
 
           await flushPromises();
@@ -523,6 +553,7 @@ describe('extension', () => {
           editExtension(provider, updateMethodMissing)(
             editorView.state,
             editorView.dispatch,
+            editorView,
           );
 
           await flushPromises();
@@ -741,6 +772,7 @@ describe('extension', () => {
       editExtension(null, Promise.resolve(extensionUpdater))(
         editorView.state,
         editorView.dispatch,
+        editorView,
       );
 
       await flushPromises();