+++ navigation/atlassian-switcher/src/common/utils/__tests__/unit/map-results-to-switcher-props.ts
@@ -713,8 +713,8 @@ describe('map-results-to-switcher-props', () => {
     });
   });
 
-  describe('mystiqueEnabled featureFlag effect on JSD', () => {
-    it('should return JSM in licensed products when FF is on', () => {
+  describe('renders JSM correctly', () => {
+    it('should return JSM in licensed products', () => {
       const props = mapResultsToSwitcherProps(
         CLOUD_ID,
         {
@@ -738,7 +738,7 @@ describe('map-results-to-switcher-props', () => {
         },
       ]);
     });
-    it('should return JSM in suggested products when FF is on', () => {
+    it('should return JSM in suggested products', () => {
       const props = mapResultsToSwitcherProps(
         CLOUD_ID,
         {
@@ -766,7 +766,7 @@ describe('map-results-to-switcher-props', () => {
         },
       ]);
     });
-    it('should return JSM in joinable sites when FF is on', () => {
+    it('should return JSM in joinable sites', () => {
       const props = mapResultsToSwitcherProps(
         CLOUD_ID,
         {