+++ navigation/atlassian-switcher/src/common/utils/map-results-to-switcher-props.ts
@@ -35,9 +35,6 @@ import { collectJoinableSiteLinks } from '../../cross-join/utils/cross-join-link
 function collectAvailableProductLinks(
   cloudId: string | null | undefined,
   mapUrl: MapUrl,
-  features: {
-    isMystiqueEnabled: boolean;
-  },
   availableProducts?: ProviderResults['availableProducts'],
 ): SwitcherItemType[] | undefined {
   if (availableProducts) {
@@ -45,12 +42,7 @@ function collectAvailableProductLinks(
       throw availableProducts.error;
     }
     if (isComplete(availableProducts)) {
-      return getAvailableProductLinks(
-        availableProducts.data,
-        cloudId,
-        mapUrl,
-        features,
-      );
+      return getAvailableProductLinks(availableProducts.data, cloudId, mapUrl);
     }
     return;
   }
@@ -210,8 +202,6 @@ export function mapResultsToSwitcherProps(
     cloudId,
     product,
   );
-  // [FD-15974]: Remove after feature flag is rolled out to 100%
-  const isMystiqueEnabled = true;
   // [FD-15975]: Remove after the FF is rolled out to 100%.
   const act959Enabled = Boolean(
     availableProducts.data?.unstableFeatures?.act959Enabled,
@@ -242,14 +232,7 @@ export function mapResultsToSwitcherProps(
   return {
     getExtendedAnalyticsAttributes,
     licensedProductLinks: collect(
-      collectAvailableProductLinks(
-        cloudId,
-        mapUrl,
-        {
-          isMystiqueEnabled,
-        },
-        availableProducts,
-      ),
+      collectAvailableProductLinks(cloudId, mapUrl, availableProducts),
       [],
     ),
     suggestedProductLinks: features.xflow
@@ -263,7 +246,6 @@ export function mapResultsToSwitcherProps(
               isDiscoverSectionEnabled: features.isDiscoverSectionEnabled,
               isDefaultEditionFreeExperimentEnabled:
                 features.isDefaultEditionFreeExperimentEnabled,
-              isMystiqueEnabled,
             },
           ),
           [],
@@ -287,10 +269,7 @@ export function mapResultsToSwitcherProps(
       ),
       [],
     ),
-    joinableSiteLinks: collect(
-      collectJoinableSiteLinks(joinableSites, { isMystiqueEnabled }),
-      [],
-    ),
+    joinableSiteLinks: collect(collectJoinableSiteLinks(joinableSites), []),
     discoverSectionLinks: hasLoadedDiscoverSection
       ? collect(
           collectDiscoverSectionLinks(