+++ navigation/atlassian-switcher/src/cross-flow/utils/cross-flow-link-collector.ts
@@ -54,7 +54,6 @@ export function collectSuggestedLinks(
   features: {
     isDiscoverSectionEnabled?: boolean;
     isDefaultEditionFreeExperimentEnabled?: boolean;
-    isMystiqueEnabled?: boolean;
   },
 ) {
   if (isError(isXFlowEnabled) || isError(provisionedProducts)) {