+++ navigation/atlassian-switcher/src/cross-join/utils/cross-join-link-collector.ts
@@ -4,15 +4,12 @@ import { ProviderResults } from '../../types';
 
 export function collectJoinableSiteLinks(
   joinableSites: ProviderResults['joinableSites'],
-  features: {
-    isMystiqueEnabled?: boolean;
-  },
 ): JoinableSiteItemType[] | undefined {
   if (joinableSites === undefined || isError(joinableSites)) {
     return [];
   }
 
   if (isComplete(joinableSites)) {
-    return getJoinableSiteLinks(joinableSites.data.sites, features);
+    return getJoinableSiteLinks(joinableSites.data.sites);
   }
 }